refactor: ManualHandlerMapping 의존성 분리 및 패키지 이동
