+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ManualHandlerMapping.java
@@ -1,0 +1,33 @@
+package com.interface21.webmvc.servlet.mvc.asis;
+
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ManualHandlerMapping implements HandlerMapping {
+
+    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
+
+    private static final Map<String, Controller> controllers = new ConcurrentHashMap<>();
+
+    @Override
+    public void initialize() {
+        log.info("Initialized Handler Mapping!");
+        controllers.keySet()
+                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
+    }
+
+    @Override
+    public Controller getHandler(final HttpServletRequest request) {
+        final var requestURI = request.getRequestURI();
+        log.debug("Request Mapping Uri : {}", requestURI);
+        return controllers.get(requestURI);
+    }
+
+    public void addController(final String path, final Controller controller) {
+        controllers.putIfAbsent(path, controller);
+    }
+}
